Rename add_location_gallery_widget.dart to dynamic_location_gallery_widget.dart
