+++ lib/ui/viewer/location/dynamic_location_gallery_widget.dart
